requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
APScheduler>=3.10.0
gunicorn>=21.2.0
rembg>=2.0.50
onnxruntime>=1.16.0
